requirements.txt
django>=1.5.1

# Additional requirements go here
django-mptt